resolves threshold to 0
